+++ resources/js/components/LoginTabLdap.vue
@@ -2,11 +2,12 @@
   <div data-test="login-tab-ldap">
     <h1 class="p-card-title">{{ props.title }}</h1>
     <form @submit.prevent="submit">
-      <div class="flex flex-col gap-2">
+      <div class="flex flex-col gap-2" data-test="username-field">
         <label :for="`${props.id}-username`">{{ props.usernameLabel }}</label>
         <InputText
           :id="`${props.id}-username`"
           v-model="username"
+          :disabled="props.loading"
           type="text"
           autocomplete="username"
           :placeholder="props.usernameLabel"
@@ -24,7 +25,7 @@
         <FormError :errors="props.errors?.username" />
       </div>
 
-      <div class="mt-6 flex flex-col gap-2">
+      <div class="mt-6 flex flex-col gap-2" data-test="password-field">
         <label :for="`${props.id}-password`">{{ props.passwordLabel }}</label>
         <Password
           v-model="password"
@@ -34,6 +35,7 @@
           toggle-mask
           required
           fluid
+          :disabled="props.loading"
           :placeholder="props.passwordLabel"
           :state="
             props.errors !== null &&