+++ CHANGELOG.md
@@ -21,12 +21,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Real-time input validation on create superuser command ([#1651])
 - Error handling in room statistics ([#1535], [#1600])
 - Error handling in room attendance ([#1535], [#1600])
+- Improve frontend tests for login page ([#1794])
 
 ### Fixed
 
 - Wrong error message shown for 422 errors when verifying email ([#1744], [#1758])
 - Broken banner link style 'warning' ([#1759], [#1760])
 - Inconsistent select/multiselect loading states ([#1772])
+- Input fields not disabled correctly on login page ([#1791], [#1794])
+- Style of 'clear' button of the room replacement selector in the 'Delete room type' dialog ([#1784], [#1787])
+- Missing form validation feedback on forgot password page ([#1802])
 
 ## [v4.2.0] - 2025-01-06
 
@@ -333,8 +337,8 @@ You can find the changelog for older versions there [here](https://github.com/TH
 [#1616]: https://github.com/THM-Health/PILOS/pull/1616
 [#1617]: https://github.com/THM-Health/PILOS/pull/1617
 [#1618]: https://github.com/THM-Health/PILOS/pull/1618
-[#1679]: https://github.com/THM-Health/PILOS/issues/1679
 [#1651]: https://github.com/THM-Health/PILOS/issues/1651
+[#1679]: https://github.com/THM-Health/PILOS/issues/1679
 [#1702]: https://github.com/THM-Health/PILOS/pull/1702
 [#1704]: https://github.com/THM-Health/PILOS/pull/1704
 [#1721]: https://github.com/THM-Health/PILOS/issues/1721
@@ -344,6 +348,11 @@ You can find the changelog for older versions there [here](https://github.com/TH
 [#1759]: https://github.com/THM-Health/PILOS/issues/1759
 [#1760]: https://github.com/THM-Health/PILOS/pull/1760
 [#1772]: https://github.com/THM-Health/PILOS/pull/1772
+[#1784]: https://github.com/THM-Health/PILOS/issues/1784
+[#1787]: https://github.com/THM-Health/PILOS/pull/1787
+[#1791]: https://github.com/THM-Health/PILOS/issues/1791
+[#1794]: https://github.com/THM-Health/PILOS/pull/1794
+[#1802]: https://github.com/THM-Health/PILOS/pull/1802
 [unreleased]: https://github.com/THM-Health/PILOS/compare/v4.2.0...develop
 [v3.0.0]: https://github.com/THM-Health/PILOS/releases/tag/v3.0.0
 [v3.0.1]: https://github.com/THM-Health/PILOS/releases/tag/v3.0.1